+++ pageserver/src/storage_sync/upload.rs
@@ -248,7 +248,7 @@ mod tests {
     use utils::lsn::Lsn;
 
     use crate::{
-        repository::repo_harness::{RepoHarness, TIMELINE_ID},
+        layered_repository::repo_harness::{RepoHarness, TIMELINE_ID},
         storage_sync::{
             index::RelativePath,
             test_utils::{create_local_timeline, dummy_metadata},
