+++ pageserver/src/walreceiver/connection_manager.rs
@@ -735,9 +735,8 @@ fn wal_stream_connection_string(
 
 #[cfg(test)]
 mod tests {
-    use crate::repository::repo_harness::{RepoHarness, TIMELINE_ID};
-
     use super::*;
+    use crate::layered_repository::repo_harness::{RepoHarness, TIMELINE_ID};
 
     #[test]
     fn no_connection_no_candidate() -> anyhow::Result<()> {
